Move to constants.js to avoid circ dependencies.
