+++ assets/js/googlesitekit/datastore/site/connection.js
@@ -26,7 +26,7 @@ import invariant from 'invariant';
  */
 import API from 'googlesitekit-api';
 import Data from 'googlesitekit-data';
-import { STORE_NAME } from './index';
+import { STORE_NAME } from './constants';
 
 const { createRegistrySelector } = Data;
