+++ assets/js/googlesitekit/datastore/site/constants.js
@@ -1,0 +1,1 @@
+export const STORE_NAME = 'core/site';
